Move add-requests functionality
