+++ app/index.js
@@ -4,7 +4,7 @@ import { join as pathJoin } from 'path'
 import pjson from '../package.json'
 import Settings from 'electron-settings'
 import './lib/report/node'
-import './setup-add-requests'
+import './add-requests'
 import rootDir from 'app-root-dir'
 import setupTrayIcon from './setup-tray-icon'
 import { report } from './lib/report/util'
